consul: Require tailscaled before restarting
Because consul was registering itself using the tailscale address, it has a
dependency on tailscale now. The change means we will wait for the tailscale
service to be online before we get the IP address from it.

The change also includes a playbook to restart the clients, which was needed
to make the cluster work again.