feat: Add ignoreDifferences option to appset
This commit adds the `ignoreDifferences` option to the `appset` library. The `ignoreDifferences` array allows users to specify which differences should be ignored when comparing Kubernetes resources. This is useful when certain fields, like secrets, should not trigger a diff. The `monitoring` appset is updated to include the `ignoreDifferences` option for the `Secret` resource, specifically the `/data/admin-password` field.